Impl ProductService

+++ src/main/java/models/Cloth.java
@@ -4,7 +4,7 @@ public class Cloth extends models.Product {
     private String size;
     private String material;
 
-    public Cloth(int id, String productName, int price, float weight, String color, int productCount, String size, String material) {
+    public Cloth(long id, String productName, float price, float weight, String color, int productCount, String size, String material) {
         super(id, productName, price, weight, color, productCount);
         this.size = size;
         this.material = material;
