+++ src/main/java/models/Product.java
@@ -1,14 +1,14 @@
 package models;
 
 public class Product {
-    protected int id;
+    protected long id;
     protected String productName;
     protected float price;
     protected float weight;
     protected String color;
     protected int productCount;
 
-    public Product(int id, String productName, int price, float weight, String color, int productCount) {
+    public Product(long id, String productName, float price, float weight, String color, int productCount) {
         this.id = id;
         this.productName = productName;
         this.price = price;
@@ -17,7 +17,7 @@ public class Product {
         this.productCount = productCount;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
